+++ hnix-store-remote/tests/NixDaemon.hs
@@ -26,8 +26,8 @@ import           System.FilePath
 import           System.Nix.Build
 import           System.Nix.StorePath
 import           System.Nix.Store.Remote
+import           System.Nix.Store.Remote.Client
 import           System.Nix.Store.Remote.GADT
-import           System.Nix.Store.Remote.Protocol hiding (WorkerOp(..))
 
 import           Crypto.Hash                    ( SHA256 )
 import           System.Nix.Nar                 ( dumpPath )
@@ -87,7 +87,7 @@ error: changing ownership of path '/run/user/1000/test-nix-store-06b0d249e561612
 
 startDaemon
   :: FilePath
-  -> IO (P.ProcessHandle, MonadStore a -> IO (Either String a, [Logger]))
+  -> IO (P.ProcessHandle, MonadStore a -> Run a)
 startDaemon fp = do
   writeConf (fp </> "etc" </> "nix.conf")
   p <- createProcessEnv fp "nix-daemon" []
@@ -108,7 +108,7 @@ enterNamespaces = do
   writeGroupMappings Nothing [GroupMapping 0 gid 1] True
 
 withNixDaemon
-  :: ((MonadStore a -> IO (Either String a, [Logger])) -> IO a) -> IO a
+  :: ((MonadStore a -> Run a) -> IO a) -> IO a
 withNixDaemon action =
   withSystemTempDirectory "test-nix-store" $ \path -> do
 
@@ -177,7 +177,10 @@ builderSh :: Text
 builderSh = "declare -xpexport > $out"
 
 spec_protocol :: Spec
-spec_protocol = Hspec.around withNixDaemon $
+spec_protocol = makeSpecProtocol withNixDaemon
+
+makeSpecProtocol :: (((MonadStore () -> Run ()) -> IO ()) -> IO ()) -> Spec
+makeSpecProtocol f = Hspec.around f $
 
   describe "store" $ do